Move selection reference helpers into selections lib
