+++ src/lib/selections.ts
@@ -21,11 +21,19 @@ import {
 } from '@src/clientSideScene/sceneConstants'
 import { AXIS_GROUP, X_AXIS } from '@src/clientSideScene/sceneUtils'
 import {
+  createCallExpressionStdLibKw,
+  createExpressionStatement,
+  createLabeledArg,
+  createLiteral,
+  nonCodeMetaEmpty,
+} from '@src/lang/create'
+import {
   findAllChildrenAndOrderByPlaceInCode,
   getEdgeCutMeta,
   getLastVariable,
   getNodeFromPath,
   getSettingsAnnotation,
+  getVariableExprsFromSelection,
   isSingleCursorInPipe,
 } from '@src/lang/queryAst'
 import { getNodePathFromSourceRange } from '@src/lang/queryAstNodePathUtils'
@@ -39,6 +47,7 @@ import { showSketchOnImportToast } from '@src/components/SketchOnImportToast'
 import { showUnsupportedSelectionToast } from '@src/components/ToastUnsupportedSelection'
 import type { KclManager } from '@src/lang/KclManager'
 import {
+  getArtifactOfTypes,
   getCapCodeRef,
   getCodeRefsByArtifactId,
   getPatternArtifactForCopyId,
@@ -56,6 +65,7 @@ import type {
   Program,
   SourceRange,
 } from '@src/lang/wasm'
+import { recast } from '@src/lang/wasm'
 import type { ArtifactEntry, ArtifactIndex } from '@src/lib/artifactIndex'
 import type {
   CommandArgument,
@@ -205,6 +215,447 @@ export async function getPrimitiveSelectionForEntity(
     parentEntityId,
     primitiveIndex: entityGetPrimitiveIndex.primitive_index,
     primitiveType: entityGetPrimitiveIndex.entity_type,
+  }
+}
+
+export type SelectionReference = {
+  id: string
+  label: string
+  code: string
+  graphSelection?: Selection
+  enginePrimitiveSelection?: EnginePrimitiveSelection
+}
+
+type ReferenceablePrimitiveSelection = EnginePrimitiveSelection & {
+  primitiveType: 'face' | 'edge'
+  graphSelection?: Selection
+  enginePrimitiveSelection?: EnginePrimitiveSelection
+}
+
+const BODY_REFERENCE_ARTIFACT_TYPES: Artifact['type'][] = [
+  'sweep',
+  'compositeSolid',
+  'pattern',
+  'helix',
+]
+
+function isReferenceablePrimitiveSelection(
+  selection: EnginePrimitiveSelection
+): selection is ReferenceablePrimitiveSelection {
+  return (
+    selection.primitiveType === 'face' || selection.primitiveType === 'edge'
+  )
+}
+
+function isBodyReferenceArtifact(
+  artifact: Artifact | undefined
+): artifact is Extract<
+  Artifact,
+  { type: 'sweep' | 'compositeSolid' | 'pattern' | 'helix' }
+> {
+  return (
+    artifact?.type === 'sweep' ||
+    artifact?.type === 'compositeSolid' ||
+    artifact?.type === 'pattern' ||
+    artifact?.type === 'helix'
+  )
+}
+
+function isSegmentReferenceArtifact(
+  artifact: Artifact | undefined
+): artifact is Extract<Artifact, { type: 'segment' }> {
+  return artifact?.type === 'segment'
+}
+
+function isPrimitiveReferenceArtifact(artifact: Artifact | undefined): boolean {
+  return (
+    artifact?.type === 'wall' ||
+    artifact?.type === 'cap' ||
+    artifact?.type === 'primitiveFace' ||
+    artifact?.type === 'sweepEdge' ||
+    artifact?.type === 'primitiveEdge' ||
+    artifact?.type === 'edgeCut'
+  )
+}
+
+function recastExpr(expr: Expr, wasmInstance: ModuleType) {
+  const code = recast(
+    {
+      start: 0,
+      end: 0,
+      moduleId: 0,
+      outerAttrs: [],
+      preComments: [],
+      commentStart: 0,
+      body: [createExpressionStatement(expr)],
+      nonCodeMeta: nonCodeMetaEmpty(),
+      shebang: null,
+      innerAttrs: [],
+    } as unknown as Parameters<typeof recast>[0],
+    wasmInstance
+  )
+  return err(code) ? null : code.trim()
+}
+
+export function getBodySelectionFromPrimitiveParentEntityId(
+  parentEntityId: string,
+  artifactGraph: ArtifactGraph,
+  {
+    bodyArtifactTypes = ['sweep', 'compositeSolid'],
+    codeRefLookup = 'last',
+    lookUpPatternCopies = false,
+  }: {
+    bodyArtifactTypes?: Artifact['type'][]
+    codeRefLookup?: 'first' | 'last'
+    lookUpPatternCopies?: boolean
+  } = {}
+): Selection | null {
+  const parentArtifact =
+    artifactGraph.get(parentEntityId) ??
+    (lookUpPatternCopies
+      ? getPatternArtifactForCopyId(parentEntityId, artifactGraph)
+      : undefined)
+  if (!parentArtifact) {
+    return null
+  }
+
+  if (
+    bodyArtifactTypes.includes(parentArtifact.type) &&
+    'codeRef' in parentArtifact
+  ) {
+    return {
+      artifact: parentArtifact,
+      codeRef: parentArtifact.codeRef,
+      engineEntityId:
+        parentArtifact.id === parentEntityId ? undefined : parentEntityId,
+    }
+  }
+
+  if (parentArtifact.type === 'path' && parentArtifact.sweepId) {
+    const parentSweep = getArtifactOfTypes(
+      { key: parentArtifact.sweepId, types: ['sweep'] },
+      artifactGraph
+    )
+    if (!err(parentSweep)) {
+      return {
+        artifact: parentSweep as Artifact,
+        codeRef: parentSweep.codeRef,
+      }
+    }
+  }
+
+  if (
+    parentArtifact.type === 'cap' ||
+    parentArtifact.type === 'wall' ||
+    parentArtifact.type === 'edgeCut'
+  ) {
+    const parentSweep = getSweepFromSuspectedSweepSurface(
+      parentArtifact.id,
+      artifactGraph
+    )
+    if (!err(parentSweep)) {
+      return {
+        artifact: parentSweep as Artifact,
+        codeRef: parentSweep.codeRef,
+      }
+    }
+  }
+
+  const parentCodeRefs = getCodeRefsByArtifactId(parentEntityId, artifactGraph)
+  if (!parentCodeRefs || parentCodeRefs.length === 0) {
+    return null
+  }
+
+  return {
+    artifact: parentArtifact,
+    codeRef:
+      codeRefLookup === 'first'
+        ? parentCodeRefs[0]
+        : parentCodeRefs[parentCodeRefs.length - 1],
+  }
+}
+
+function createPrimitiveReferenceCode({
+  primitiveSelection,
+  artifactGraph,
+  kclManager,
+  wasmInstance,
+}: {
+  primitiveSelection: ReferenceablePrimitiveSelection
+  artifactGraph: ArtifactGraph
+  kclManager: KclManager
+  wasmInstance: ModuleType
+}): string | null {
+  if (!primitiveSelection.parentEntityId) {
+    return null
+  }
+
+  const bodySelection = getBodySelectionFromPrimitiveParentEntityId(
+    primitiveSelection.parentEntityId,
+    artifactGraph,
+    {
+      bodyArtifactTypes: BODY_REFERENCE_ARTIFACT_TYPES,
+      codeRefLookup: 'first',
+      lookUpPatternCopies: true,
+    }
+  )
+  if (!bodySelection) {
+    return null
+  }
+
+  const bodyVariables = getVariableExprsFromSelection(
+    { graphSelections: [bodySelection], otherSelections: [] },
+    artifactGraph,
+    kclManager.ast,
+    wasmInstance,
+    undefined,
+    {
+      lastChildLookup: true,
+      artifactTypeFilter: BODY_REFERENCE_ARTIFACT_TYPES,
+    }
+  )
+  if (err(bodyVariables) || bodyVariables.exprs.length === 0) {
+    return null
+  }
+
+  const bodyExpr = bodyVariables.exprs[0]
+  const functionName =
+    primitiveSelection.primitiveType === 'face' ? 'faceId' : 'edgeId'
+  const call = createCallExpressionStdLibKw(
+    functionName,
+    structuredClone(bodyExpr),
+    [
+      createLabeledArg(
+        'index',
+        createLiteral(primitiveSelection.primitiveIndex, wasmInstance)
+      ),
+    ]
+  )
+
+  return recastExpr(call, wasmInstance)
+}
+
+function createExpressionReferences({
+  label,
+  selection,
+  artifactGraph,
+  kclManager,
+  wasmInstance,
+  options,
+}: {
+  label: string
+  selection: Selection
+  artifactGraph: ArtifactGraph
+  kclManager: KclManager
+  wasmInstance: ModuleType
+  options?: Parameters<typeof getVariableExprsFromSelection>[5]
+}): SelectionReference[] {
+  const variableExprs = getVariableExprsFromSelection(
+    { graphSelections: [selection], otherSelections: [] },
+    artifactGraph,
+    kclManager.ast,
+    wasmInstance,
+    undefined,
+    options
+  )
+  if (err(variableExprs)) {
+    return []
+  }
+
+  return variableExprs.exprs.flatMap((expr) => {
+    const code = recastExpr(expr, wasmInstance)
+    if (!code) {
+      return []
+    }
+
+    return [
+      {
+        id: `${label}:${selection.artifact?.id || selection.engineEntityId || code}:${code}`,
+        label,
+        code,
+        graphSelection: selection,
+      },
+    ]
+  })
+}
+
+export async function getSelectionReferences({
+  graphSelections,
+  enginePrimitives,
+  artifactGraph,
+  engineCommandManager,
+  kclManager,
+  wasmInstance,
+}: {
+  graphSelections: Selection[]
+  enginePrimitives: EnginePrimitiveSelection[]
+  artifactGraph: ArtifactGraph
+  engineCommandManager: ConnectionManager
+  kclManager: KclManager
+  wasmInstance: ModuleType
+}): Promise<SelectionReference[]> {
+  const references: SelectionReference[] = []
+  const primitiveSelections: ReferenceablePrimitiveSelection[] = []
+
+  for (const selection of graphSelections) {
+    if (isBodyReferenceArtifact(selection.artifact)) {
+      references.push(
+        ...createExpressionReferences({
+          label: 'Body',
+          selection,
+          artifactGraph,
+          kclManager,
+          wasmInstance,
+          options: {
+            lastChildLookup: true,
+            artifactTypeFilter: BODY_REFERENCE_ARTIFACT_TYPES,
+          },
+        })
+      )
+      continue
+    }
+
+    if (isSegmentReferenceArtifact(selection.artifact)) {
+      references.push(
+        ...createExpressionReferences({
+          label: 'Segment',
+          selection,
+          artifactGraph,
+          kclManager,
+          wasmInstance,
+        })
+      )
+      continue
+    }
+
+    if (!isPrimitiveReferenceArtifact(selection.artifact)) {
+      continue
+    }
+
+    const entityId = selection.artifact?.id || selection.engineEntityId
+    if (!entityId) {
+      continue
+    }
+
+    const primitiveSelection = await getPrimitiveSelectionForEntity(
+      entityId,
+      engineCommandManager
+    )
+    if (
+      primitiveSelection &&
+      isReferenceablePrimitiveSelection(primitiveSelection)
+    ) {
+      primitiveSelections.push({
+        ...primitiveSelection,
+        graphSelection: selection,
+      })
+    }
+  }
+
+  for (const selection of enginePrimitives) {
+    if (isReferenceablePrimitiveSelection(selection)) {
+      primitiveSelections.push({
+        ...selection,
+        enginePrimitiveSelection: selection,
+      })
+    }
+  }
+
+  const dedupedPrimitiveSelections = [
+    ...new Map(
+      primitiveSelections.map((selection) => [
+        `${selection.primitiveType}:${selection.parentEntityId || ''}:${selection.primitiveIndex}:${selection.entityId}`,
+        selection,
+      ])
+    ).values(),
+  ]
+
+  references.push(
+    ...dedupedPrimitiveSelections.flatMap((primitiveSelection) => {
+      const code = createPrimitiveReferenceCode({
+        primitiveSelection,
+        artifactGraph,
+        kclManager,
+        wasmInstance,
+      })
+      if (!code) {
+        return []
+      }
+
+      return [
+        {
+          id: `${primitiveSelection.primitiveType}:${primitiveSelection.entityId}`,
+          label: primitiveSelection.primitiveType === 'face' ? 'Face' : 'Edge',
+          code,
+          graphSelection: primitiveSelection.graphSelection,
+          enginePrimitiveSelection: primitiveSelection.enginePrimitiveSelection,
+        },
+      ]
+    })
+  )
+
+  return [
+    ...new Map(
+      references.map((reference) => [
+        `${reference.label}:${reference.code}`,
+        reference,
+      ])
+    ).values(),
+  ]
+}
+
+function isSameCodeRange(left: Selection, right: Selection) {
+  return (
+    left.codeRef.range[0] === right.codeRef.range[0] &&
+    left.codeRef.range[1] === right.codeRef.range[1]
+  )
+}
+
+function isSameGraphSelection(left: Selection, right: Selection) {
+  if (left.artifact?.id && right.artifact?.id) {
+    return left.artifact.id === right.artifact.id
+  }
+
+  if (left.engineEntityId && right.engineEntityId) {
+    return left.engineEntityId === right.engineEntityId
+  }
+
+  return isSameCodeRange(left, right)
+}
+
+function isSameEnginePrimitiveSelection(
+  left: EnginePrimitiveSelection,
+  right: EnginePrimitiveSelection
+) {
+  return left.entityId === right.entityId
+}
+
+export function removeReferenceFromSelections(
+  selections: Selections,
+  reference: SelectionReference
+): Selections {
+  const graphSelectionToRemove = reference.graphSelection
+  const enginePrimitiveSelectionToRemove = reference.enginePrimitiveSelection
+
+  return {
+    graphSelections: graphSelectionToRemove
+      ? selections.graphSelections.filter(
+          (selection) =>
+            !isSameGraphSelection(selection, graphSelectionToRemove)
+        )
+      : selections.graphSelections,
+    otherSelections: enginePrimitiveSelectionToRemove
+      ? selections.otherSelections.filter(
+          (selection) =>
+            !(
+              isEnginePrimitiveSelection(selection) &&
+              isSameEnginePrimitiveSelection(
+                selection,
+                enginePrimitiveSelectionToRemove
+              )
+            )
+        )
+      : selections.otherSelections,
   }
 }
